Move RootActivity to foundation package
