+++ app/src/main/res/layout/activity_root.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.RootActivity">
+        tools:context=".view.foundation.RootActivity">
 
     <fragment
             android:name="androidx.navigation.fragment.NavHostFragment"
